[Fix] app/linesフォルダの記載をmodel側に移行する
 - models/以下に以下のファイルを生成。
   ・cat_line_bot.rb：同ファイルをクラスメソッドに修正
   ・scheduler.rb：lib以下にあったタスクをクラスメソッドに修正
 - models/conerns以下に以下のファイルを生成。
   ・message_event.rb：同ファイルを移行修正。
